requirements.txt
amqp==2.6.1
asgiref==3.3.4
autopep8==1.5.7
billiard==3.6.4.0
celery==4.4.1
click==7.1.2
click-didyoumean==0.0.3
click-plugins==1.1.1
click-repl==0.1.6
Django==3.2
django-debug-toolbar==3.2.1
djangorestframework==3.12.4
djangorestframework-jwt==1.11.0
kombu==4.6.11
Markdown==3.3.4
Pillow==8.2.0
prompt-toolkit==3.0.18
pycodestyle==2.7.0
PyJWT==1.7.1
pytz==2021.1
redis==3.5.3
six==1.16.0
sqlparse==0.4.1
toml==0.10.2
vine==1.3.0
wcwidth==0.2.5
